Automatic theme determination at first start
Determines at the first start which color should be used.

That means if in Windows/Mac/Linux you determine that a dark theme should be used, the default is set to True and reversed.

For existing users this does not change anything. But users who start Runebook for the first time will get the right layout directly.